+++ src/types/complex-types/object.ts
@@ -21,7 +21,7 @@ import { ComplexType } from "./complex-type"
 import { getPrimitiveFactoryFromValue } from "../primitives"
 import { optional } from "../utility-types/optional"
 import { isReferenceFactory } from "../utility-types/reference"
-import { isIdentifierFactory } from "../utility-types/identifier"
+import { isIdentifierFactory, IIdentifierDescriptor } from "../utility-types/identifier"
 import { Property } from "../property-types/property"
 import { IdentifierProperty } from "../property-types/identifier-property"
 import { ReferenceProperty } from "../property-types/reference-property"
@@ -106,7 +106,7 @@ export class ObjectType extends ComplexType<any, any> {
             } else if (isIdentifierFactory(value)) {
                 invariant(!this.identifierAttribute, `Cannot define property '${key}' as object identifier, property '${this.identifierAttribute}' is already defined as identifier property`)
                 this.identifierAttribute = key
-                this.props[key] = new IdentifierProperty(key)
+                this.props[key] = new IdentifierProperty(key, (value as IIdentifierDescriptor<any>).primitiveType)
             } else if (isPrimitive(value)) {
                 const baseType = getPrimitiveFactoryFromValue(value)
                 this.props[key] = new ValueProperty(key, optional(baseType, value))