+++ src/types/property-types/identifier-property.ts
@@ -2,10 +2,14 @@ import { getMSTAdministration } from "../../core"
 import { extendObservable, observable, IObjectWillChange } from "mobx"
 import { Property } from "./property"
 import { isValidIdentifier, fail } from "../../utils"
+import { IType, typecheck } from "../type"
 
 export class IdentifierProperty extends Property {
-    constructor(propertyName: string) {
+    subtype: IType<any, any>
+
+    constructor(propertyName: string, subtype: IType<any, any>) {
         super(propertyName)
+        this.subtype = subtype
     }
 
     initialize(target: any, snapshot: any) {
@@ -15,21 +19,24 @@ export class IdentifierProperty extends Property {
     }
 
     willChange(change: IObjectWillChange): IObjectWillChange | null {
-        if (!isValidIdentifier(change.newValue))
-            fail(`Not a valid identifier: '${change.newValue}`)
+        const identifier = change.newValue
+        if (typeof identifier !== "number" && !isValidIdentifier(identifier))
+            fail(`Not a valid identifier: '${identifier}`)
+        typecheck(this.subtype, identifier)
         const node = getMSTAdministration(change.object)
         node.assertWritable()
         const oldValue = change.object[this.name]
-        if (oldValue !== undefined && oldValue !== change.newValue)
-            fail(`It is not allowed to change the identifier of an object, got: '${change.newValue}' but expected: '${oldValue}'`)
+        if (oldValue !== undefined && oldValue !== identifier)
+            fail(`It is not allowed to change the identifier of an object, got: '${identifier}' but expected: '${oldValue}'`)
 
         return change
     }
 
     serialize(instance: any, snapshot: any) {
-        if (!isValidIdentifier(instance[this.name]))
-            fail(`Object does not have a valid identifier yet: '${instance[this.name]}'`)
-        snapshot[this.name] = instance[this.name]
+        const identifier = instance[this.name]
+        if (!this.isValidIdentifier(identifier))
+            fail(`Object does not have a valid identifier yet: '${identifier}'`)
+        snapshot[this.name] = identifier
     }
 
     deserialize(instance: any, snapshot: any) {
@@ -37,6 +44,12 @@ export class IdentifierProperty extends Property {
     }
 
     isValidSnapshot(snapshot: any) {
-        return isValidIdentifier(snapshot[this.name])
+        return this.isValidIdentifier(snapshot[this.name])
+    }
+
+    isValidIdentifier(identifier: any): boolean {
+        if (typeof identifier !== "number" && !isValidIdentifier(identifier))
+            return false
+        return this.subtype.is(identifier)
     }
 }